+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-Cluj56i5.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-BZRMTYlD.js";import{_ as a}from"./mermaid.core-D0LM0ocJ.js";import"./chunk-5VM5RSS4-UU7MnpKr.js";import"./chunk-XXDRQBXY-DQeV_gtI.js";import"./chunk-POPQ4Y6H-CmbTE3e0.js";import"./chunk-F27PBJKO-D7lavHAd.js";import"./channel-CGvebdaW.js";import"./index-Dq4h0gqk.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};