+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-OXjC16PD.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-C_rngRBV.js";import{_ as a}from"./mermaid.core-By_UUvbT.js";import"./chunk-5VM5RSS4-CbXO_XXy.js";import"./chunk-XXDRQBXY-DWHptcG1.js";import"./chunk-POPQ4Y6H-D3iDRQ02.js";import"./chunk-F27PBJKO-BsvFJURD.js";import"./channel-Y2MAGI6m.js";import"./index-CYVHwe7m.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};